requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.10
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect2==0.7.0
defusedxml==0.8.0rc2
deprecation==2.1.0
Django==5.1.4
django-celery==3.1.17
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-email-verification==0.3.3
django-environ==0.12.0
django-htmx==1.21.0
django-mathfilters==1.0.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.8
flower==2.0.1
fonttools==4.55.6
humanize==4.11.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg-binary==3.2.4
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sorl-thumbnail==12.11.0
sqlparse==0.5.3
stripe==11.4.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
